==> src/notebooks/1-prep_coltivazioni.ipynb <==
Questo notebook è orientato alla preparazione dei dati relativi alle coltivazioni. Contiene diverse celle di codice che eseguono operazioni come la lettura di file CSV, la pulizia e la manipolazione dei dati, e infine il salvataggio dei dati puliti in un nuovo file CSV. 

Importazione delle librerie necessarie

# Importa le librerie necessarie
import pandas as pd

Lettura del file CSV


# Lettura dataset da un file CSV . Dataset scaricato dal sito istat 
frumento = pd.read_excel("../../data/raw/ita/coltivazioni_istat.xlsx")

Esplorazione iniziale del DataFrame

#Vediamo il dataset per comprenderne la struttura e il contenuto 
display(frumento)

Operazioni sui dati 

# Ridenominazione colonne e filtraggio delle colonne/righe non corrette
# Allo stesso temnpo si compongono delle stringhe che saranno i nomi finali delle colonne

# Inizializza il numero della colonna e l'anno iniziale
num = 2
anno = 2006

# Nomi iniziali delle colonne basati sulle prime colonne
col_name = f'{anno}'
col_name_next = f'Unnamed: {num}' if num+2 < len(frumento.columns) else None

# Combina anno e tipo di dato per i nomi delle colonne
a = col_name
b = frumento.loc[0, col_name] if col_name else None
c = frumento.loc[0, col_name_next] if col_name_next else None

# Aggiorna il DataFrame con i nomi delle colonne combinati
frumento.loc[1, col_name] = str(a) + ' - ' + str(b)
frumento.loc[1, col_name_next] = str(a) + ' - ' + str(c)
num+=2
anno += 1

# Loop attraverso le restanti colonne per costruire e impostare i nuovi nomi delle colonne
while num < len(frumento.columns):        
    col_name = f'{anno}'
    col_name_next = f'Unnamed: {num}' if num+1 <= len(frumento.columns) else None

    a = col_name
    b = frumento.loc[0, col_name] if col_name else None
    c = frumento.loc[0, col_name_next] if col_name_next else None
    frumento.loc[1, col_name] = str(a) + ' - ' + str(b)
    frumento.loc[1, col_name_next] = str(a) + ' - ' + str(c)
    num+=2
    anno+=1

display(frumento)

# Si impostano nuovi nomi per tutte le colonne
frumento.columns = frumento.iloc[1].tolist()
display(frumento)

# Elimina le prime due righe che contengono metadati
frumento.drop(0, inplace=True)
frumento.drop(1, inplace=True)

# Rinomina la colonna 'Territorio' in 'zona'
frumento.rename(columns={'Territorio': 'zona'}, inplace=True)

# Imposta 'zona' come indice del DataFrame
frumento.set_index("zona")

# Typo fix e salvataggio
frumento = frumento.map(lambda x: x.lower() if isinstance(x, str) else x)
display(frumento)

Salvataggio dataset modificato

# Salva il DataFrame pulito in un file CSV
frumento.to_csv(f"../../data/raw/ita/dataset_frumento_prov_ita.csv", index=False)
