==> src/notebooks/2-prep_temp_prec.ipynb <==
Questo notebook si concentra sulla preparazione dei dati relativi alla temperatura e alle precipitazioni

Importazione delle librerie necessarie

# Importa le librerie necessarie
import pandas as pd
import numpy as np


# Lettura dataset
prov = pd.read_excel("../../data/raw/ita/temp_prec.xlsx")

# Rinomina la colonna 'COMUNI' in 'provincia'
prov.rename(columns={'COMUNI': 'provincia'}, inplace=True)
display(prov)

try:
      # Tenta di eliminare la colonna 'Unnamed: 17' se esiste
	prov.drop('Unnamed: 17', axis=1, inplace=True)
except KeyError:
    # Ignora l'errore se la colonna non esiste
    pass

Separazione dei dati sulla temperatura

# Copia delle colonne realtive alle temperature medie
temp = prov.iloc[:, :17].copy()
temp.iloc[0, 0] = 'provincia'

# Rinominazione colonne
temp_mapping = dict(zip(temp.columns, temp.iloc[0].values))  
temp.rename(columns=temp_mapping, inplace=True)
temp = temp.drop(0)

# Aggiunta del suffisso ' - temp cel' ai nomi delle colonne
new_cols = [str(col) + ' - temp cel' for col in temp.columns[1:]]
new_cols.insert(0, 'provincia')
temp.columns = new_cols

# Conversione dei nomi delle province in minuscolo
temp['provincia'] = temp['provincia'].str.lower()

# Visualizza il DataFrame delle temperature
display(temp)
# Salva il DataFrame delle temperature in un file CSV
temp.to_csv("../../data/raw/ita/temperature2006-2021.csv", index=False)

# Copia delle colonne realtive alle precipitazioni
prec = prov.iloc[:, 17:].copy() 

# Si copia la colonna con le province e si pone all'inizio
prec['provincia'] = temp.iloc[:, 0]

ultima_colonna = prec.pop(prec.columns[-1])
prec.insert(0, ultima_colonna.name, ultima_colonna)
prec.iloc[0, 0] = 'provincia'

# Rinominazione colonne
prec_mapping = dict(zip(prec.columns, prec.iloc[0].values))
prec.rename(columns=prec_mapping, inplace=True)
prec = prec.drop(0)

# Aggiunta del suffisso ' - prec mm' ai nomi delle colonne
new_cols = [str(col) + ' - prec mm' for col in prec.columns[1:]]
new_cols.insert(0, 'provincia')
prec.columns = new_cols

# Conversione dei nomi delle province in minuscolo
prec['provincia'] = prec['provincia'].str.lower()

# Visualizza il DataFrame delle precipitazioni
display(prec)
prec.to_csv("../../data/raw/ita/precipitazioni2006-2021.csv", index=False)
