==> src/notebooks/3-prep_geo.ipynb <==
Questo notebook si concentra sulla preparazione dei dati geografici.

Importazione delle librerie necessarie

# Importa le librerie necessarie
import pandas as pd
import numpy as np

Lettura del file CSV


# Lettura file CSV contenente i dati di altitudine dei comuni
alt = pd.read_excel("../../data/raw/ita/comuni_alt.xlsx")
display(alt)

# Lettura del file CSV contenente le coordinate longitudine e latitudine dei comuni
lonlat = pd.read_excel("../../data/raw/ita/comuni_lon_lat.xlsx")
display(lonlat)

# Rimuove le righe con valori nulli nel dataset delle coordinate
lonlat = lonlat.dropna()

# Rimuove la colonna 'istat' e ordina il dataset per nome del comune
lonlat2 = lonlat.drop("istat", axis=1)
lonlat2.sort_values(by='comune', inplace=True)
display(lonlat2)

# Rimuove le righe con valori nulli nel dataset delle altitudini
alt = alt.dropna()

# Calcola la media tra 'ALT_MIN' e 'ALT_MAX' per ottenere l'altitudine media e la aggiunge come nuova colonna 'alt'
alt['alt'] = alt[['ALT_MIN', 'ALT_MAX']].mean(axis=1)

# Rimuove le colonne non necessarie dal dataset delle altitudini
alt2 = alt.drop(["COD_REG", "COD_PRO", "PRO_COM", "SUPERFICIE (in KMQ)", "ALT_MIN", "ALT_MAX", "RANGE", "MEDIA", "MEDIANA", "STD"], axis=1)
display(alt2)

# Si effettua il merge dei due dataset per ottenerne uno che contiene ogni comune con rispettive coordinate
# Merge dei dataset delle altitudini e delle coordinate, unendo i dati basandosi sul nome del comune
comuni = pd.merge(alt2, lonlat2, left_on='NOME', right_on='comune')

# Rimuove la colonna duplicata 'comune' e ordina il dataset per nome del comune
comuni = comuni.drop("comune", axis=1)
comuni.sort_values(by='NOME', inplace=True)

# Ordina le colonne in ordine alfabetico
comuni = comuni.sort_index(axis=1)

# Rinomina alcune colonne per maggiore chiarezza
nuovi_nomi = {'NOME': 'comune', 'lng': 'lon'}

comuni = comuni.rename(columns=nuovi_nomi)
display(comuni)

# Lettura dataset comuni e province
com = pd.read_excel("../../data/raw/ita/elenco_comuni.xlsx")
prov = pd.read_excel("../../data/raw/ita/elenco_province.xlsx")

# Eliminazione colonne non necessarie dal dataset dei comuni
comuni2 = com.drop(["codice_istat", "denominazione_ita_altra", "denominazione_altra", "flag_capoluogo", "codice_belfiore", "lat", "lon", "superficie_kmq", "codice_sovracomunale"], axis=1)

# Rimuove le colonne non necessarie dal dataset delle province
prov2 = prov.drop(["codice_regione", "tipologia_provincia", "numero_comuni", "superficie_kmq", "codice_sovracomunale"], axis=1)

# Merge dataset per ottenerne uno con comuni e rispettive province
comuni_prov = pd.merge(comuni2, prov2, on='sigla_provincia')

# Ordina il dataset per nome del comune e rimuove le righe con valori nulli
comuni_prov.sort_values(by='denominazione_ita', inplace=True)
comuni_prov.dropna(inplace=True)

# Rimuove la colonna 'sigla_provincia' e rinomina le colonne per maggiore chiarezza
comuni_prov.drop("sigla_provincia", axis=1, inplace=True)
nuovi_nomi = {'denominazione_ita': 'comune', 'denominazione_provincia': 'provincia'}
comuni_prov = comuni_prov.rename(columns=nuovi_nomi)
display(comuni_prov)

# Merge finale, il dataset finale ha comune, provincia e dati geografici
finale = pd.merge(comuni, comuni_prov, on='comune')

# Riordina le colonne del dataset finale per avere un layout coerente
finale = finale.reindex(columns=["comune", "provincia", "alt", "lat", "lon"])
display(finale)

# Salva il dataset finale in un file CSV
finale.to_csv("../../data/raw/ita/comuni_geo.csv", index=False)
