==> src/notebooks/4-final.ipynb <==
Questo notebook si concentra sulla creazione di un dataset finale combinando dati geografici, superfici, produzioni, precipitazioni e temperature.

Importazione delle librerie necessarie

import pandas as pd
import numpy as np

Lettura dei file CSV

# Lettura dataset comuni italiani e dati delle coltivazioni di frumento 
frumento = pd.read_csv("../../data/raw/ita/dataset_frumento_prov_ita.csv")
comuni = pd.read_csv("../../data/raw/ita/comuni_geo.csv")

# Conversione del nome della provincia in minuscolo per facilitare il merge
comuni['provincia'] = comuni['provincia'].str.lower()

# Visualizzazione dei dataset letti
display(frumento)
display(comuni)

# Merge in base alla provincia
# Combina i dati dei comuni con quelli della coltivazione di frumento sulla base della provincia
merge = pd.merge(comuni, frumento, left_on='provincia', right_on='zona')

# Si rimuovono colonne per le quali non ci sono dati
# Rimuove le colonne non necessarie o con troppi valori mancanti
merge = merge.drop(['zona', '2022 - superficie totale - ettari', '2022 - produzione totale - quintali',	'2023 - superficie totale - ettari', '2023 - produzione totale - quintali'], axis = 1)

# Si uniscono anche i dati relativi alle precipitazioni e temperature
# Lettura dei dataset relativi alle precipitazioni e alle temperature
prec = pd.read_csv('../../data/raw/ita/precipitazioni2006-2021.csv')
temp = pd.read_csv('../../data/raw/ita/temperature2006-2021.csv')

# Merge finale
# Combina i dati delle precipitazioni e delle temperature con il dataset principale
merge = merge.merge(prec, on='provincia').merge(temp, on='provincia')

# I NaN sono indicati con .. e ...., li sostituisco con np.nan per comodità
# Sostituisce i valori mancanti indicati con ".." o "...." con np.nan per una gestione più semplice dei NaN
merge.replace(['..', '....'], np.nan, inplace=True)
display(merge)

# Calcolo delle statistiche sui valori NaN (decommentare per eseguire)
# Calcola e stampa la percentuale di valori mancanti per ogni colonna
# totale = 0
# for colonna in merge.columns:
#     numero_valori_nulli = merge[colonna].isna().sum()
#     totale += numero_valori_nulli
#     numero_totale_valori = len(merge[colonna])
#     percentuale_valori_nulli = (numero_valori_nulli / numero_totale_valori) * 100

#     print(f"{colonna} --> {percentuale_valori_nulli:.2f}%")
    
# print("Numero totale di Nan: " + str(totale))
# print("Percentuale totale Nan: " + str(totale/merge.size*100))

# # Decommenta per calcolo delle statistiche, quanti valori NaN per ogni riga
# totale = 0
# righe_con_null = 0
# for indice_riga, riga in merge.iterrows():
#     numero_valori_nulli = riga.isna().sum()
#     if numero_valori_nulli > 0:
#         righe_con_null += 1
#     totale += numero_valori_nulli
#     numero_totale_valori = len(riga)
#     percentuale_valori_nulli = (numero_valori_nulli / numero_totale_valori) * 100

#     print(f"{percentuale_valori_nulli:.2f}%")

# print("Numero totale di Nan: " + str(totale))
# print("Percentuale totale Nan: " + str(totale/merge.size*100))
# print("Righe con valori Nan: " + str(righe_con_null))

# Converto i dati numerici del dataframe in interi
# Converte i dati numerici delle colonne in interi, forzando il tipo numerico e gestendo gli errori con coerce
merge.iloc[:, 5:] = merge.iloc[:, 5:].apply(pd.to_numeric, errors='coerce')

# Sostituisco i NaN con la mediana della rispettiva colonna
merge.iloc[:, 5:] = merge.iloc[:, 5:].fillna(merge.iloc[:, 5:].median())

display(merge)

#### Operazioni finali sul dataset
Si effettuano ulteriori operazioni sul dataset per ottenerne due versioni finali:
* La prima, `dataset_final.csv` \
Si modifica il dataset per avere su un'unica riga: dati del comune, dati geografici, anno di riferimento, dati coltivazioni, dati meteorologici. \
Features: comune, provincia, alt, lat, lon, anno, superficie totale - ettari, produzione totale - quintali, prec mm, temp cel
* La seconda, `features.csv` \
A partire dal dataframe che origina `dataset_final.csv`, si calcolano le feature in maniera aggregata, così da avere un dataset riassuntivo, andando ad eliminare la feature `anno`. \
Ogni riga contiene dati del comune, dati geografici, media e deviazione standard dei dati di superficie, produzione, precipitazioni e temperatura. \
Feature: comune, provincia, alt, lat, lon, superficie_media, produzione_media, prec_media, temp_media, superficie_std, produzione_std, prec_std, temp_std

# Rimuove eventuali spazi nei nomi delle colonne
merge.columns = [col.strip() for col in merge.columns]

# Identifica le colonne da unire
id_vars = ['comune', 'provincia', 'alt', 'lat', 'lon']
value_vars_surface = [col for col in merge.columns if 'superficie totale - ettari' in col]
value_vars_production = [col for col in merge.columns if 'produzione totale - quintali' in col]
value_vars_prec = [col for col in merge.columns if 'prec mm' in col]
value_vars_temp = [col for col in merge.columns if 'temp cel' in col]

# Unisce per superficie totale
# Trasforma i dati di superficie totale da wide a long format
surface_long = pd.melt(merge, id_vars=id_vars, value_vars=value_vars_surface,
                       var_name='anno', value_name='superficie totale - ettari')
surface_long['anno'] = surface_long['anno'].str.split(' - ').str[0].astype(int)

# Unisce per produzione totale
# Trasforma i dati di produzione totale da wide a long format
production_long = pd.melt(merge, id_vars=id_vars, value_vars=value_vars_production,
                          var_name='anno', value_name='produzione totale - quintali')
production_long['anno'] = production_long['anno'].str.split(' - ').str[0].astype(int)

# Unisce per precipitazioni
# Trasforma i dati di precipitazioni da wide a long format
prec_long = pd.melt(merge, id_vars=id_vars, value_vars=value_vars_prec,
                    var_name='anno', value_name='prec mm')
prec_long['anno'] = prec_long['anno'].str.split(' - ').str[0].astype(int)

# Unisce per temperatura
# Trasforma i dati di temperatura da wide a long format
temp_long = pd.melt(merge, id_vars=id_vars, value_vars=value_vars_temp,
                    var_name='anno', value_name='temp cel')
temp_long['anno'] = temp_long['anno'].str.split(' - ').str[0].astype(int)

# Unisce i dataframe lunghi
# Combina tutti i dataframe long in un unico dataframe finale
merge_long = surface_long.merge(production_long, on=id_vars + ['anno'])
merge_long = merge_long.merge(prec_long, on=id_vars + ['anno'])
merge_long = merge_long.merge(temp_long, on=id_vars + ['anno'])

# Raggruppa i dati in base alle colonne ['comune', 'provincia', 'alt', 'lat', 'lon'].
# Questo significa che per ogni combinazione di questi valori, verranno calcolate le statistiche di aggregazione.
# Per ogni comune, vengono calcolate:
# superficie_media: la media della colonna 'superficie totale - ettari'.
# produzione_media: la media della colonna 'produzione totale - quintali'.
# prec_media: la media della colonna 'prec mm' (precipitazioni in millimetri).
# temp_media: la media della colonna 'temp cel' (temperatura in gradi Celsius).
# superficie_std: la deviazione standard della colonna 'superficie totale - ettari'.
# produzione_std: la deviazione standard della colonna 'produzione totale - quintali'.
# prec_std: la deviazione standard della colonna 'prec mm'.
# temp_std: la deviazione standard della colonna 'temp cel'.

features = merge_long.groupby(['comune', 'provincia', 'alt', 'lat', 'lon']).agg(
    superficie_media=('superficie totale - ettari', 'mean'),
    produzione_media=('produzione totale - quintali', 'mean'),
    prec_media=('prec mm', 'mean'),
    temp_media=('temp cel', 'mean'),
    superficie_std=('superficie totale - ettari', 'std'),
    produzione_std=('produzione totale - quintali', 'std'),
    prec_std=('prec mm', 'std'),
    temp_std=('temp cel', 'std'),
).reset_index()

# Visualizzazione dei dataframe finali
display(merge_long)
display(features)

# Si ottiene un dataset completo per ogni comune si hanno i dati geografici, superfici, produzioni, precipitazioni e temperature per ogni anno
merge_long.to_csv("../../data/processed/dataset_final.csv")

# Si aggregano i dati, così da avere una riga per comune
features.to_csv("../../data/processed/features.csv")
