==> src/notebooks/4c-final_india.ipynb <==
import pandas as pd

df = pd.read_csv("../../data/raw/india/india_crop_production.csv")
display(df)

df = df[(df['Crop'] == 'Wheat') & (df['Crop_Year'].between(2006, 2023))]
df = df.drop(['Season', 'Crop'], axis=1)
df.rename(columns={'State_Name': 'stato', 'District_Name': 'citta', 'Crop_Year': 'anno', 'Area': 'superficie', 'Production': 'produzione'}, inplace=True)
df = df.applymap(lambda x: x.lower() if isinstance(x, str) else x)

display(df)

cities = df.groupby(['stato', 'citta']).agg(
    superficie_media=('superficie', 'mean'),
    produzione_media=('produzione', 'mean'),
    superficie_std=('superficie', 'std'),
    produzione_std=('produzione', 'std')
).reset_index()

# Visualizza i primi risultati
cities = cities.dropna()

display(cities)

geo = pd.read_csv("../../data/raw/india/indian_cities_geolocations.csv", encoding='ISO-8859-1')
display(geo)

geo = geo.drop(['Geoname ID', 'Name', 'Alternate Names', 'Population', 'Country', 'Coordinates'], axis=1)
geo.rename(columns={'ASCII Name': 'citta', 'Latitude': 'lat', 'Longitude': 'lon', 'Digital Elevation Model': 'alt'}, inplace=True)
geo = geo.applymap(lambda x: x.lower() if isinstance(x, str) else x)
geo['alt'] = geo['alt'].astype('float64')
display(geo)

india = pd.merge(cities, geo, on='citta')
india.dropna()

# Nuovo ordine delle colonne
new_column_order = ['stato', 'citta', 'alt', 'lat', 'lon', 'superficie_media', 'produzione_media', 'superficie_std', 'produzione_std']

# Riordinare il DataFrame
india = india[new_column_order]

display(india)

temp = pd.read_csv("../../data/raw/india/temperature_india.csv")
display(temp)

# Escludere la prima colonna e calcolare la media delle colonne adiacenti
# In questo caso, calcoliamo la media di _MaxTemp e _MinTemp
temp['jan'] = temp[['JAN_MaxTemp', 'JAN_MinTemp']].mean(axis=1)
temp['feb'] = temp[['FEB_MaxTemp', 'FEB_MinTemp']].mean(axis=1)
temp['mar'] = temp[['MAR_MaxTemp', 'MAR_MinTemp']].mean(axis=1)
temp['apr'] = temp[['APR_MaxTemp', 'APR_MinTemp']].mean(axis=1)
temp['may'] = temp[['MAY_MaxTemp', 'MAY_MinTemp']].mean(axis=1)
temp['jun'] = temp[['JUN_MaxTemp', 'JUN_MinTemp']].mean(axis=1)
temp['jul'] = temp[['JUL_MaxTemp', 'JUL_MinTemp']].mean(axis=1)
temp['aug'] = temp[['AUG_MaxTemp', 'AUG_MinTemp']].mean(axis=1)
temp['sep'] = temp[['SEP_MaxTemp', 'SEP_MinTemp']].mean(axis=1)
temp['oct'] = temp[['OCT_MaxTemp', 'OCT_MinTemp']].mean(axis=1)
temp['nov'] = temp[['NOV_MaxTemp', 'NOV_MinTemp']].mean(axis=1)
temp['dec'] = temp[['DEC_MaxTemp', 'DEC_MinTemp']].mean(axis=1)
temp = temp.drop(['JAN_MaxTemp', 'JAN_MinTemp', 'FEB_MaxTemp', 'FEB_MinTemp', 'MAR_MaxTemp', 'MAR_MinTemp', 'APR_MaxTemp', 'APR_MinTemp', 'MAY_MaxTemp', 'MAY_MinTemp', 'JUN_MaxTemp', 'JUN_MinTemp', 'JUL_MaxTemp', 'JUL_MinTemp', 'AUG_MaxTemp', 'AUG_MinTemp', 'SEP_MaxTemp', 'SEP_MinTemp', 'OCT_MaxTemp', 'OCT_MinTemp', 'NOV_MaxTemp', 'NOV_MinTemp', 'DEC_MaxTemp', 'DEC_MinTemp'], axis=1)

temp['temp_media'] = temp.iloc[:, 1:13].mean(axis=1)

temp = temp.drop(['jan', 'feb', 'mar', 'apr', 'may', 'jun', 'jul', 'aug', 'sep', 'oct', 'nov', 'dec'], axis=1)
temp.rename(columns={'location': 'citta'}, inplace=True)

temp = temp.map(lambda x: x.lower() if isinstance(x, str) else x)

display(temp)

india_temp = pd.merge(india, temp, on='citta', how='left')
# india_temp.dropna()
india_temp['temp_media'] = india_temp['temp_media'].fillna(india_temp['temp_media'].mean())
india_temp['produzione_media'] = india_temp['produzione_media'].fillna(india_temp['produzione_media'].mean())

# Nuovo ordine delle colonne
new_column_order = ['stato', 'citta', 'alt', 'lat', 'lon', 'superficie_media', 'produzione_media', 'superficie_std', 'produzione_std', 'temp_media']

# Riordinare il DataFrame
india_temp = india_temp[new_column_order]

display(india_temp)

df_rain = pd.read_csv('../../data/raw/india/india_rainfall.csv')

# Filtra per gli anni tra il 2006 e il 2021
df_rain_fil = df_rain[(df_rain['YEAR'] >= 2006) & (df_rain['YEAR'] <= 2023)]

# Seleziona solo le colonne SUBDIVISION, YEAR e ANNUAL
selected_columns_df = df_rain_fil[['SUBDIVISION', 'YEAR', 'ANNUAL']]
display(selected_columns_df)

rainfall = selected_columns_df.groupby(['SUBDIVISION']).agg(
    prec_media=('ANNUAL', 'mean')
).reset_index()

rainfall.rename(columns={'SUBDIVISION': 'stato'}, inplace=True)
rainfall = rainfall.map(lambda x: x.lower() if isinstance(x, str) else x)
rainfall['stato'] = rainfall['stato'].str.replace('&', 'and')

# Visualizza i primi risultati
rainfall = rainfall.dropna()

display(rainfall)

india_final = pd.merge(rainfall, india_temp, on='stato')

# Nuovo ordine delle colonne
new_column_order = ['stato', 'citta', 'alt', 'lat', 'lon', 'superficie_media', 'produzione_media', 'superficie_std', 'produzione_std', 'temp_media', 'prec_media']

# Riordinare il DataFrame
india_final = india_final[new_column_order]

india_final.to_csv("../../data/processed/features_india.csv")

display(india_final)
